\title{snpMatrix-internal}
\name{snpMatrix-internal}
\alias{genotype-class}
\alias{haplotype-class}
\alias{convert.snpMatrix}
\alias{convert.snpMatrix.dir}
\description{All the dirty details that doesn't belong elsewhere. At the
  moment just for hiding references to the \code{genotype-class} and
  \code{haplotype-class} class which is in the genetics package.
}
\keyword{package}
