\title{snpMatrix-internal}
\name{snpMatrix-internal}
\alias{genotype-class}
\alias{haplotype-class}
\description{All the dirty details that doesn't belong elsewhere. At the
  moment just for hiding references to the \code{genotype-class} and
  \code{haplotype-class} class which is in the genetics package. The two
  links should be \link[genetics]{genotype-class} and
  \link[genetics]{haplotype-class} .
}
\keyword{package}
